Organise FE into pages/ and rename legacy *Original components
Move route-level components out of src/components into a dedicated
src/pages folder so pages and reusable components are clearly separated,
and drop the leftover "Original" suffixes now that the code is current.

- src/pages: HomePage, HelpPage, PvpLobbyPage, AllShipLayoutsPage,
  SinglePlayerPage (was AppOriginal/AiPage), PvpGamePage (was GamePage)
- GameOriginal -> SinglePlayerGame, GridOriginal -> SinglePlayerBoard
- Update imports in main.tsx, the moved files, and the test; fix stale
  paths in PandaBattleship.FE.esproj

Route components now uniformly use the *Page suffix. Type check and the
full vitest suite pass.
