+++ PandaBattleship.FE/src/pages/AllShipLayoutsPage.tsx
@@ -1,5 +1,5 @@
-import DisplayAllLayouts from "./DisplayAllLayouts";
-import { PageHeader } from "./PageHeader";
+import DisplayAllLayouts from "../components/DisplayAllLayouts";
+import { PageHeader } from "../components/PageHeader";
 
 export const AllShipLayoutsPage: React.FC = () => (
     <div className="max-w-5xl mx-auto p-4 text-center min-h-screen flex flex-col items-center">
